Complete Windows Unicode and LLP64 audit fixes

+++ src/cluster_asm.c
@@ -2403,9 +2403,10 @@ int slotSnapshotSaveRio(int req, rio *rdb, int *error) {
                     if (!send_slot_info) {
                         /* Format slot info */
                         char buf[128];
-                        int len = snprintf(buf, sizeof(buf), "%d:%lu:%lu",
-                                    k, kvstoreDictSize(db->keys, k),
-                                    kvstoreDictSize(db->expires, k));
+                        int len = snprintf(buf, sizeof(buf), "%d:%llu:%llu",
+                                    k,
+                                    (unsigned long long)kvstoreDictSize(db->keys, k),
+                                    (unsigned long long)kvstoreDictSize(db->expires, k));
                         serverAssert(len > 0 && len < (int)sizeof(buf));
 
                         /* Send slot info */
