+++ src/rio.c
@@ -605,7 +605,7 @@ uint8_t rioCheckType(rio *r) {
  * generating the Redis protocol for the Append Only File. */
 
 /* Write multi bulk count in the format: "*<count>\r\n". */
-size_t rioWriteBulkCount(rio *r, char prefix, long count) {
+size_t rioWriteBulkCount(rio *r, char prefix, long long count) {
     char cbuf[128];
     int clen;
