+++ src/script_lua.h
@@ -62,7 +62,7 @@ void* luaGetFromRegistry(lua_State* lua, const char* name);
 void luaCallFunction(scriptRunCtx* r_ctx, lua_State *lua, robj** keys, size_t nkeys, robj** args, size_t nargs, int debug_enabled);
 void luaExtractErrorInformation(lua_State *lua, errorInfo *err_info);
 void luaErrorInformationDiscard(errorInfo *err_info);
-unsigned long luaMemory(lua_State *lua);
+size_t luaMemory(lua_State *lua);
 void luaGC(lua_State *lua, int *gc_count);
 
 #endif /* __SCRIPT_LUA_H_ */
